reduced complexity in public api and tests. all tests pass
