+++ public_api/database/tests/mock_fast_api_testclient.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 
-from database_client.db_client_interface import DBClientInterface
+from interfaces.db_client_interface import DBClientInterface
 
 
 class MockDBClient(DBClientInterface):
@@ -10,5 +10,5 @@ class MockDBClient(DBClientInterface):
     def get(self, url, stream=False):
         return self.client.get(url=url)
 
-    def post(self, url, params, files):
+    def post(self, url, params=None, files=None):
         return self.client.post(url=url, params=params, files=files)
