+++ public_api/database/tests/test_db_requests_impl.py
@@ -1,20 +1,26 @@
 import os
-import random
-import tempfile
+import secrets
 import unittest
 
-import dotenv
+os.environ["TZ"] = "Europe/Copenhagen"
+os.environ["API_HOSTNAME"] = "private_api"
+os.environ["API_PORT"] = "7000"
+os.environ["API_URL"] = "http://private_api:7000"
+os.environ["LOG_LEVEL"] = "10"
+
+os.environ["DATA_DIR"] = "/opt/app/data/"
+os.environ["API_TASKS"] = "/api/tasks/"
+
+os.environ["API_OUTPUT_ZIPS"] = "/api/tasks/outputs/"
+os.environ["API_INPUT_ZIPS"] = "/api/tasks/inputs/"
+os.environ["API_MODELS"] = "/api/models/"
+os.environ["API_MODELS_BY_HUMAN_READABLE_ID"] = "/api/models/human_readable_id/"
+os.environ["API_MODEL_ZIPS"] = "/api/models/zips/"
 
 from database.db_impl import DBImpl
-from testing.mock_components.mock_db import MockDB
-from testing.mock_components.mock_fast_api_testclient import MockDBClient
+from database.tests.mock_fast_api_testclient import MockDBClient
 from testing.mock_components.mock_models_and_tasks import MockModelsAndTasks
-from testing.mock_components.mock_private_fast_api import MockPrivateFastAPI
-
-from exceptions.exceptions import NoZipAttachedException
-
-dotenv.load_dotenv()
-
+from database.tests.mock_private_fast_api import MockPrivateFastAPI
 
 class TestDBImpl(unittest.TestCase):
     """
@@ -24,19 +30,20 @@ class TestDBImpl(unittest.TestCase):
     def setUp(self) -> None:
         self.base_dir = ".tmp"
         self.repo = MockModelsAndTasks()
-        self.db_backend = MockDB()
-        self.app = MockPrivateFastAPI(db=self.db_backend)
-        self.db_client = MockDBClient(self.app)
+        self.private_api = MockPrivateFastAPI()
+
+        self.db_client = MockDBClient(self.private_api)
         self.db = DBImpl(db_client=self.db_client)
 
     def tearDown(self):
         self.repo.purge()
-        self.db_backend.purge()
+        self.private_api.purge()
 
     def test_post_task_intended(self):
         with open(self.repo.task.input_zip, "br") as r:
             echo = self.db.post_task(zip_file=r,
-                                     model_human_readable_id=self.repo.model.human_readable_id)
+                                     model_human_readable_id=self.repo.model.human_readable_id,
+                                     uid=secrets.token_urlsafe())
         self.assertIsNotNone(echo)
         self.assertEqual(echo["model_human_readable_id"], self.repo.model.human_readable_id)
         self.assertIn("uid", echo.keys())
@@ -47,10 +54,7 @@ class TestDBImpl(unittest.TestCase):
         with open(self.repo.model_zip, "br") as r:
             model = self.db.post_model(container_tag=self.repo.model.container_tag,
                                        human_readable_id=self.repo.model.human_readable_id,
-                                       input_mountpoint=self.repo.model.input_mountpoint,
-                                       output_mountpoint=self.repo.model.output_mountpoint,
                                        zip_file=r,
-                                       model_mountpoint=self.repo.model.model_mountpoint,
                                        description=self.repo.model.description,
                                        model_available=self.repo.model.model_available,
                                        use_gpu=self.repo.model.use_gpu)
