+++ public_api/interfaces/db_interface.py
@@ -1,32 +1,26 @@
-import tempfile
 from abc import abstractmethod
 from typing import Optional, BinaryIO
-
-import requests
 
 
 class DBInterface:
     @abstractmethod
     def post_task(self, model_human_readable_id: str,
                   zip_file: BinaryIO,
-                  uid=None):
+                  uid: str):
         pass
 
     @abstractmethod
-    def get_output_zip_by_uid(self, uid: str):
+    def get_output_zip(self, uid: str):
         pass
 
     @abstractmethod
-    def delete_task_by_uid(self, uid: str):
+    def delete_task(self, uid: str):
         pass
 
     @abstractmethod
     def post_model(self,
                    container_tag: str,
                    human_readable_id: str,
-                   input_mountpoint: Optional[str] = None,
-                   output_mountpoint: Optional[str] = None,
-                   model_mountpoint: Optional[str] = None,
                    description: Optional[str] = None,
                    zip_file: Optional[BinaryIO] = None,
                    model_available: Optional[bool] = True,
