+++ public_api/interfaces/public_api_interface.py
@@ -1,10 +1,11 @@
-import logging
 from abc import abstractmethod
-from typing import Optional, Any, Dict
+from typing import Optional, Any
 
 from fastapi import File, UploadFile, FastAPI
-from fastapi.responses import StreamingResponse, Response
-from database.db_interface import DBInterface
+from fastapi.responses import StreamingResponse
+
+from interfaces.db_interface import DBInterface
+
 
 class PublicFastAPIInterface(FastAPI):
     @abstractmethod
@@ -20,29 +21,26 @@ class PublicFastAPIInterface(FastAPI):
 
         @abstractmethod
         def public_post_task(model_human_readable_id: str,
-                      zip_file: UploadFile = File(...),
-                      uid=None) -> str:
+                             zip_file: UploadFile = File(...)
+                             ) -> str:
             pass
 
         @abstractmethod
-        def public_get_output_zip_by_uid(uid: str) -> StreamingResponse:
+        def public_get_output_zip(uid: str) -> StreamingResponse:
             pass
 
         @abstractmethod
-        def public_delete_task_by_uid(uid: str):
+        def public_delete_task(uid: str):
             pass
 
         @abstractmethod
         def public_post_model(container_tag: str,
-                       human_readable_id: str,
-                       input_mountpoint: Optional[str] = None,
-                       output_mountpoint: Optional[str] = None,
-                       model_mountpoint: Optional[str] = None,
-                       description: Optional[str] = None,
-                       zip_file: Optional[UploadFile] = File(None),
-                       model_available: Optional[bool] = True,
-                       use_gpu: Optional[bool] = True,
-                       ):
+                              human_readable_id: str,
+                              description: Optional[str] = None,
+                              zip_file: Optional[UploadFile] = File(None),
+                              model_available: Optional[bool] = True,
+                              use_gpu: Optional[bool] = True,
+                              ):
             pass
 
         @abstractmethod
