+++ public_api/public_api/public_api_fast_api_impl.py
@@ -5,18 +5,16 @@ import secrets
 import tempfile
 import threading
 import traceback
-from typing import Any, Optional, Union, Dict
+from typing import Any, Optional, Union
 from urllib.parse import urljoin
 
 import requests
-from fastapi import FastAPI, UploadFile, File, HTTPException
-from fastapi.responses import StreamingResponse, Response
+from fastapi import FastAPI, UploadFile, HTTPException
+from fastapi.responses import StreamingResponse
+from interfaces.public_api_interface import PublicFastAPIInterface
 
-from database.db_interface import DBInterface
-from exceptions.exceptions import PostTaskException, ZipFileMissingException, ZipFileShouldBeNoneException, TaskOutputZipNotFound
-from public_api.public_api_interface import PublicFastAPIInterface
-
-app = FastAPI()
+from exceptions.exceptions import ZipFileMissingException, ZipFileShouldBeNoneException
+from interfaces.db_interface import DBInterface
 
 threads = []
 
@@ -33,30 +31,37 @@ class PublicFastAPI(PublicFastAPIInterface):
         def public_hello_world():
             return {"message": "Hello world - Welcome to the public database API"}
 
-        @self.post(os.environ['PUBLIC_POST_TASK'])
+        @self.post(os.environ['API_TASKS'])
         def public_post_task(model_human_readable_id: str,
                              zip_file: UploadFile,
                              ) -> str:
+            def post_task_thread(uid, model_human_readable_id, zip_file):
+                logging.info(f"[ ] Posting task: {uid} on {model_human_readable_id}")
+                try:
+                    res = self.db.post_task(model_human_readable_id=model_human_readable_id,
+                                        zip_file=zip_file,
+                                        uid=uid)
+                except Exception:
+                    traceback.print_exc()
+                finally:
+                    zip_file.close()
 
             models = self.db.get_models()
             if not model_human_readable_id in [model.human_readable_id for model in models]:
                 raise HTTPException(550, detail="ModelNotFound")
 
-            def post_task_thread(uid, model_human_readable_id, zip_file, ):
-                logging.info(f"[ ] Posting task: {uid} on {model_human_readable_id}")
-                return self.db.post_task(model_human_readable_id=model_human_readable_id,
-                                         zip_file=zip_file,
-                                         uid=uid)
-
-            uid = secrets.token_urlsafe(32)  # Give this request a unique identifie}
+            uid = secrets.token_urlsafe()  # Give this request a unique identifie}
+            tmp_zip_file = tempfile.TemporaryFile()
+            tmp_zip_file.write(zip_file.file.read())
+            tmp_zip_file.seek(0)
             t = threading.Thread(target=post_task_thread, kwargs={"uid": uid,
                                                                   "model_human_readable_id": model_human_readable_id,
-                                                                  "zip_file": zip_file.file})
+                                                                  "zip_file": tmp_zip_file})
             t.start()
             self.threads.append(t)
             return uid
 
-        @self.get(urljoin(os.environ['PUBLIC_GET_OUTPUT_ZIP_BY_UID'], "{uid}"))
+        @self.get(urljoin(os.environ['API_OUTPUT_ZIPS'], "{uid}"))
         def public_get_output_zip_by_uid(uid: str) -> StreamingResponse:
             def iterfile(bytes_from_db: bytes):
                 with tempfile.TemporaryFile() as tmp_file:
@@ -64,7 +69,7 @@ class PublicFastAPI(PublicFastAPIInterface):
                     tmp_file.seek(0)
                     yield from tmp_file
 
-            res = self.db.get_output_zip_by_uid(uid)
+            res = self.db.get_output_zip(uid)
 
             if res.ok:
                 return StreamingResponse(iterfile(bytes_from_db=res.content))
@@ -74,9 +79,9 @@ class PublicFastAPI(PublicFastAPIInterface):
                 except requests.HTTPError:
                     raise HTTPException(status_code=res.status_code, detail=json.loads(res.content))
 
-        @self.delete(urljoin(os.environ['PUBLIC_GET_OUTPUT_ZIP_BY_UID'], "{uid}"))
+        @self.delete(urljoin(os.environ['API_OUTPUT_ZIPS'], "{uid}"))
         def public_delete_task_by_uid(uid: str):
-            status = self.db.delete_task_by_uid(uid)
+            status = self.db.delete_task(uid)
 
             try:
                 return status
@@ -84,7 +89,7 @@ class PublicFastAPI(PublicFastAPIInterface):
                 traceback.print_exc()
                 raise HTTPException(status_code=550, detail=str(e))
 
-        @self.get(os.environ["PUBLIC_GET_MODELS"])
+        @self.get(os.environ["API_MODELS"])
         def public_get_models():
             try:
                 return self.db.get_models()
@@ -92,12 +97,9 @@ class PublicFastAPI(PublicFastAPIInterface):
                 raise HTTPException(status_code=550, detail=str(e))
 
         if bool(os.environ.get("ALLOW_PUBLIC_POST_MODEL")):
-            @self.post(os.environ.get("PUBLIC_POST_MODEL"))
+            @self.post(os.environ.get("API_MODELS"))
             def public_post_model(container_tag: str,
                                   human_readable_id: str,
-                                  input_mountpoint: Union[str, None] = None,
-                                  output_mountpoint: Union[str, None] = None,
-                                  model_mountpoint: Union[str, None] = None,
                                   description: Union[str, None] = None,
                                   model_available: Union[bool, None] = None,
                                   use_gpu: Union[bool, None] = None,
@@ -115,9 +117,6 @@ class PublicFastAPI(PublicFastAPIInterface):
                 try:
                     return self.db.post_model(container_tag=container_tag,
                                               human_readable_id=human_readable_id,
-                                              input_mountpoint=input_mountpoint,
-                                              output_mountpoint=output_mountpoint,
-                                              model_mountpoint=model_mountpoint,
                                               description=description,
                                               zip_file=zip_file,
                                               model_available=model_available,
